Violation Disposition bar chart
Here we see that the vast majority of cases have been settled (paid) by the violators. Surprisingly, the second most common outcome is "Barred by CPLR", which means that the statute of limitations has been reach and the matter can no longer be pursued. Unsurprisingly, the fourth most common outcome is a default on the claim (defendant failed to appear in court or formally respond).